changing robots and maps to be comparable
robots of same size, maps of same size, laser of same size and angle ( now in degree in bullet) .
Swap left and right wheel in bullet.
Changed the speed in bullet according to the real robot max velocity (500mm/s).
Following camera  now in the yml file (desactivate for better performance) Same as the GUI option.